Autorizacao Parte1

+++ app/Http/Controllers/ProjectController.php
@@ -30,7 +30,7 @@ class ProjectController extends Controller
      */
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id'=>\Authorizer::getResourceOwnerId()]);
     }
 
 
@@ -53,6 +53,11 @@ class ProjectController extends Controller
      */
     public function show($id)
     {
+
+        if($this->checkProjectOwner($id)==false){
+            return ['error'=>'sem permissão'];
+        }
+
         return $this->repository->find($id);
     }
 
@@ -66,7 +71,9 @@ class ProjectController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        if($this->checkProjectOwner($id)==false){
+        return ['error'=>'sem permissão'];
+    }
 
         return $this->service->update($request->all(),$id);
     }
@@ -78,6 +85,17 @@ class ProjectController extends Controller
      */
     public function destroy($id)
     {
+        if($this->checkProjectOwner($id)==false){
+            return ['error'=>'sem permissão'];
+        }
        return $this->repository->delete($id);
+    }
+
+    private function checkProjectOwner($projectId){
+        $userId = \Authorizer::getResourceOwnerId();
+
+       return $this->repository->isOwner($projectId, $userId);
+
+
     }
 }
